pyproject.toml
[project]
name = "st-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
    "pillow-heif>=0.20.0",
    "streamlit-list-widget>=0.1.11",
    "loguru>=0.7.2",
]
